# backend/src/main/java/com/campaignorganizer/campaign/adapter/beatkind/in/web/BeatKindController.java
package com.campaignorganizer.campaign.adapter.beatkind.in.web;

import com.campaignorganizer.campaign.adapter.beatkind.in.web.BeatKindWebDtos.BeatKindRequest;
import com.campaignorganizer.campaign.adapter.beatkind.in.web.BeatKindWebDtos.BeatKindResponse;
import com.campaignorganizer.campaign.application.beatkind.port.in.CreateBeatKindUseCase;
import com.campaignorganizer.campaign.application.beatkind.port.in.DeleteBeatKindUseCase;
import com.campaignorganizer.campaign.application.beatkind.port.in.GetBeatKindUseCase;
import com.campaignorganizer.campaign.application.beatkind.port.in.ListBeatKindsUseCase;
import com.campaignorganizer.campaign.application.beatkind.port.in.UpdateBeatKindUseCase;
import jakarta.validation.Valid;
import java.net.URI;
import java.util.List;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/worlds/{worldId}/beat-kinds")
public class BeatKindController {

    private final CreateBeatKindUseCase createUseCase;
    private final UpdateBeatKindUseCase updateUseCase;
    private final DeleteBeatKindUseCase deleteUseCase;
    private final GetBeatKindUseCase getUseCase;
    private final ListBeatKindsUseCase listUseCase;
    private final BeatKindWebMapper mapper;

    public BeatKindController(CreateBeatKindUseCase createUseCase, UpdateBeatKindUseCase updateUseCase,
                              DeleteBeatKindUseCase deleteUseCase, GetBeatKindUseCase getUseCase,
                              ListBeatKindsUseCase listUseCase, BeatKindWebMapper mapper) {
        this.createUseCase = createUseCase;
        this.updateUseCase = updateUseCase;
        this.deleteUseCase = deleteUseCase;
        this.getUseCase = getUseCase;
        this.listUseCase = listUseCase;
        this.mapper = mapper;
    }

    @GetMapping
    public List<BeatKindResponse> list(@PathVariable UUID worldId) {
        return listUseCase.list(worldId).stream().map(mapper::toResponse).toList();
    }

    @GetMapping("/{beatKindId}")
    public BeatKindResponse get(@PathVariable UUID worldId, @PathVariable UUID beatKindId) {
        return mapper.toResponse(getUseCase.get(worldId, beatKindId));
    }

    @PostMapping
    public ResponseEntity<BeatKindResponse> create(@PathVariable UUID worldId,
                                                    @Valid @RequestBody BeatKindRequest request) {
        BeatKindResponse response =
                mapper.toResponse(createUseCase.create(mapper.toCreateCommand(worldId, request)));
        return ResponseEntity
                .created(URI.create("/api/worlds/" + worldId + "/beat-kinds/" + response.id()))
                .body(response);
    }

    @PutMapping("/{beatKindId}")
    public BeatKindResponse update(@PathVariable UUID worldId, @PathVariable UUID beatKindId,
                                   @Valid @RequestBody BeatKindRequest request) {
        return mapper.toResponse(
                updateUseCase.update(mapper.toUpdateCommand(worldId, beatKindId, request)));
    }

    @DeleteMapping("/{beatKindId}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@PathVariable UUID worldId, @PathVariable UUID beatKindId) {
        deleteUseCase.delete(worldId, beatKindId);
    }
}