# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Campaign Organizer API
  version: 0.1.0
  description: >
    Personal worldbuilding and RPG campaign management application.
    This document is the **canonical, design-first contract** for the API.
    Both the Spring Boot backend and the React/TypeScript frontend are expected
    to conform to it (see docs/adr/0008-contract-first-openapi.md).

    Phase 0 covers single-password authentication and the `worlds` resource,
    which is the top-level container every other domain object will hang off of.
  license:
    name: Proprietary (personal use)
  contact:
    name: Campaign Organizer
servers:
  - url: /api
    description: Same-origin API root (frontend proxies /api to the backend)
...
paths:
...
  /worlds:
    get:
      tags: [worlds]
      summary: List all worlds
      operationId: listWorlds
      responses:
        '200':
          description: All worlds owned by the user, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/World'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [worlds]
      summary: Create a world
      operationId: createWorld
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorldRequest'
      responses:
        '201':
          description: World created.
          headers:
            Location:
              description: URI of the created world.
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/World'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /worlds/{worldId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
    get:
      tags: [worlds]
      summary: Get a single world
      operationId: getWorld
      responses:
        '200':
          description: The world.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/World'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [worlds]
      summary: Update a world
      operationId: updateWorld
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorldRequest'
      responses:
        '200':
          description: World updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/World'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [worlds]
      summary: Delete a world
      operationId: deleteWorld
      responses:
        '204':
          description: World deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /worlds/{worldId}/beat-kinds:
    parameters:
      - $ref: '#/components/parameters/WorldId'
    get:
      tags: [beat-kinds]
      summary: List a world's beat kinds
      operationId: listBeatKinds
      responses:
        '200':
          description: Beat kinds in the world.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BeatKind'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [beat-kinds]
      summary: Create a beat kind in the world's catalog
      operationId: createBeatKind
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeatKindRequest'
      responses:
        '201':
          description: Beat kind created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeatKind'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A beat kind with that name already exists in this world.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /worlds/{worldId}/beat-kinds/{beatKindId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
      - $ref: '#/components/parameters/BeatKindId'
    get:
      tags: [beat-kinds]
      summary: Get a beat kind
      operationId: getBeatKind
      responses:
        '200':
          description: The beat kind.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeatKind'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [beat-kinds]
      summary: Rename or recolor a beat kind
      operationId: updateBeatKind
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeatKindRequest'
      responses:
        '200':
          description: Beat kind updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BeatKind'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A beat kind with that name already exists in this world.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
    delete:
      tags: [beat-kinds]
      summary: Delete a beat kind (un-tags any beats that used it)
      operationId: deleteBeatKind
      responses:
        '204':
          description: Beat kind deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /worlds/{worldId}/handouts/{handoutId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
      - $ref: '#/components/parameters/HandoutId'
    get:
      tags: [handouts]
      summary: Get a handout
      operationId: getHandout
      responses:
        '200':
          description: The handout.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Handout'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [handouts]
      summary: Update a handout
      operationId: updateHandout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoutRequest'
      responses:
        '200':
          description: Handout updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Handout'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [handouts]
      summary: Delete a handout
      operationId: deleteHandout
      responses:
        '204':
          description: Handout deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /worlds/{worldId}/handout-categories:
    parameters:
      - $ref: '#/components/parameters/WorldId'
    get:
      tags: [handouts]
      summary: List a world's handout categories (ADR-0105)
      operationId: listHandoutCategories
      responses:
        '200':
          description: Handout categories in the world, ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HandoutCategory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [handouts]
      summary: Create a handout category
      operationId: createHandoutCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoutCategoryRequest'
      responses:
        '201':
          description: Handout category created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandoutCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /worlds/{worldId}/handout-categories/{categoryId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
      - $ref: '#/components/parameters/CategoryId'
    put:
      tags: [handouts]
      summary: Update a handout category
      operationId: updateHandoutCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HandoutCategoryRequest'
      responses:
        '200':
          description: Handout category updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HandoutCategory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [handouts]
      summary: Delete a handout category (its handouts are kept, uncategorized)
      operationId: deleteHandoutCategory
      responses:
        '204':
          description: Handout category deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /worlds/{worldId}/campaigns/{campaignId}/arcs/{arcId}/beats/{beatId}:
    parameters:
      - $ref: '#/components/parameters/WorldId'
      - $ref: '#/components/parameters/CampaignId'
      - $ref: '#/components/parameters/ArcId'
      - $ref: '#/components/parameters/BeatId'
    put:
      tags: [arcs]
      summary: Update a beat
      operationId: updateBeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BeatRequest'
      responses:
        '200':
          description: Beat updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Beat'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [arcs]
      summary: Delete a beat
      operationId: deleteBeat
      responses:
        '204':
          description: Beat deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Bearer token obtained from POST /auth/login. Signed with HS256.
  parameters:
    WorldId:
      name: worldId
      in: path
      required: true
      description: Identifier of the world.
      schema:
        type: string
        format: uuid
    CategoryId:
      name: categoryId
      in: path
      required: true
      description: Identifier of the category.
      schema:
        type: string
        format: uuid
...
    CampaignId:
      name: campaignId
      in: path
      required: true
      description: Identifier of the campaign.
      schema:
        type: string
        format: uuid
...
    BeatKindId:
      name: beatKindId
      in: path
      required: true
      description: Identifier of the beat kind.
      schema:
        type: string
        format: uuid
...
    HandoutId:
      name: handoutId
      in: path
      required: true
      description: Identifier of the handout.
      schema:
        type: string
        format: uuid
    ArcId:
      name: arcId
      in: path
      required: true
      description: Identifier of the arc.
      schema:
        type: string
        format: uuid
...
    BeatId:
      name: beatId
      in: path
      required: true
      description: Identifier of the beat.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Missing or invalid bearer token / password.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    BeatRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: [string, "null"]
          maxLength: 20000
          description: Beat notes (Markdown, see ADR-0054).
        done:
          type: boolean
          default: false
        articleIds:
          type: array
          items:
            type: string
            format: uuid
          description: Linked articles (place, NPC, item, …).
        statblockIds:
          type: array
          items:
            type: string
            format: uuid
          description: Linked statblocks (goblin spearman, boss, …).
        encounterIds:
          type: array
          items:
            type: string
            format: uuid
          description: Linked encounters (ADR-0097) — structured, quantified combat groupings.
        tableIds:
          type: array
          items:
            type: string
            format: uuid
          description: Linked roll tables.
        deckIds:
          type: array
          items:
            type: string
            format: uuid
          description: Linked card decks.
        sessionId:
          type: [string, "null"]
          format: uuid
          description: Optional session where this beat occurred.
        kindId:
          type: [string, "null"]
          format: uuid
          description: Optional GM-defined beat kind tag (ADR-0101), informational only.
        position:
          type: [integer, "null"]
    Beat:
      type: object
      required: [id, arcId, title, done, articleIds, statblockIds, encounterIds, tableIds, deckIds, position, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        arcId:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: [string, "null"]
          description: Beat notes (Markdown, see ADR-0054).
        done:
          type: boolean
        articleIds:
          type: array
          items:
            type: string
            format: uuid
        statblockIds:
          type: array
          items:
            type: string
            format: uuid
        encounterIds:
          type: array
          items:
            type: string
            format: uuid
        tableIds:
          type: array
          items:
            type: string
            format: uuid
        deckIds:
          type: array
          items:
            type: string
            format: uuid
        sessionId:
          type: [string, "null"]
          format: uuid
        kindId:
          type: [string, "null"]
          format: uuid
          description: Optional GM-defined beat kind tag (ADR-0101), informational only.
        position:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BeatKindRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        color:
          type: [string, "null"]
          maxLength: 20
    BeatKind:
      type: object
      required: [id, worldId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        worldId:
          type: string
          format: uuid
        name:
          type: string
        color:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    HandoutRequest:
      type: object
      required: [title, preset]
      properties:
        categoryId:
          type: [string, "null"]
          format: uuid
          description: Handout category (ADR-0105); null for uncategorized.
        title:
          type: string
          minLength: 1
          maxLength: 200
        preset:
          type: string
          enum: [PARCHMENT, NEWSPAPER, POSTER, LETTER]
          description: Visual style applied to the printed handout.
        body:
          type: [string, "null"]
          description: Markdown body (wiki-links render as anchors).
        sessionId:
          type: [string, "null"]
          format: uuid
          description: >
            Session this handout is prepped for (ADR-0077); null for a
            general/unassigned handout. Operational grouping for packet
            assembly only - not a wiki-graph link (ADR-0070).
        revealed:
          type: boolean
          default: false
          description: >
            Screen-only prep flag: has the GM revealed this to players?
            Not access control (single-user app, no player-facing surface)
            and doesn't affect printing.
...
    Handout:
      type: object
      required: [id, worldId, title, preset, body, revealed, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        worldId:
          type: string
          format: uuid
        categoryId:
          type: [string, "null"]
          format: uuid
        title:
          type: string
        preset:
          type: string
          enum: [PARCHMENT, NEWSPAPER, POSTER, LETTER]
        body:
          type: [string, "null"]
        sessionId:
          type: [string, "null"]
          format: uuid
        revealed:
          type: boolean
          description: Screen-only prep flag; not access control (ADR-0079).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    HandoutCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        parentId:
          type: [string, "null"]
          format: uuid
          description: Parent handout category for nesting, or null for a root category.
    HandoutCategory:
      type: object
      required: [id, worldId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        worldId:
          type: string
          format: uuid
        parentId:
          type: [string, "null"]
          format: uuid
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    WorldRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          examples: ["Aethelgard"]
        description:
          type: string
          maxLength: 5000
          description: Optional summary of the world.
        scratch:
          type: boolean
          default: false
          description: >-
            Cosmetic-only sandbox/brainstorming flag (FR-60) — no functional
            effect, just a label in the world list and switcher.
    World:
      type: object
      required: [id, name, scratch, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: [string, "null"]
        layerStyles:
          $ref: '#/components/schemas/LayerStyleMap'
        scratch:
          type: boolean
          description: >-
            Cosmetic-only sandbox/brainstorming flag (FR-60) — no functional
            effect, just a label in the world list and switcher.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LayerStyle:
      type: object
      description: Styling for a map layer.
      properties:
        color:
          type: [string, "null"]
        icon:
          type: [string, "null"]
    LayerStyleMap:
      type: object
      description: Per-layer styling keyed by layer name.
      additionalProperties:
        $ref: '#/components/schemas/LayerStyle'
...
    ProblemDetail:
      type: object
      description: RFC 9457 / RFC 7807 problem response.
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
security:
  - bearerAuth: []